// packages/localnet-ui/src/explorer/cell-inspector/block.tlb.generated.ts
export interface Bool {
    readonly kind: 'Bool';
    readonly value: boolean;
}
// ...
export type Maybe<X> = Maybe_nothing<X> | Maybe_just<X>;
// ...
export interface ExtraCurrencyCollection {
    readonly kind: 'ExtraCurrencyCollection';
    readonly dict: Dictionary<number, bigint>;
}

/*
currencies$_ grams:Grams other:ExtraCurrencyCollection
           = CurrencyCollection;
*/

export interface CurrencyCollection {
    readonly kind: 'CurrencyCollection';
    readonly grams: bigint;
    readonly other: ExtraCurrencyCollection;
}
// ...
export interface InMsgDescr {
    readonly kind: 'InMsgDescr';
    readonly anon0: Dictionary<bigint, {value: InMsg, extra: ImportFees}>;
}
// ...
export interface OutMsgDescr {
    readonly kind: 'OutMsgDescr';
    readonly anon0: Dictionary<bigint, {value: OutMsg, extra: CurrencyCollection}>;
}
// ...
export interface ShardAccountBlocks {
    readonly kind: 'ShardAccountBlocks';
    readonly anon0: Dictionary<bigint, {value: AccountBlock, extra: CurrencyCollection}>;
}
// ...
export interface ShardIdent {
    readonly kind: 'ShardIdent';
    readonly shard_pfx_bits: number;
    readonly workchain_id: number;
    readonly shard_prefix: bigint;
}
// ...
export interface BlkMasterInfo {
    readonly kind: 'BlkMasterInfo';
    readonly master: ExtBlkRef;
}
// ...
export interface BlockInfo {
    readonly kind: 'BlockInfo';
    readonly prev_seq_no?: number;
    readonly version: number;
    readonly not_master: number;
    readonly after_merge: number;
    readonly before_split: number;
    readonly after_split: number;
    readonly want_split: Bool;
    readonly want_merge: Bool;
    readonly key_block: Bool;
    readonly vert_seqno_incr: number;
    readonly flags: number;
    readonly seq_no: number;
    readonly vert_seq_no: number;
    readonly shard: ShardIdent;
    readonly gen_utime: number;
    readonly start_lt: bigint;
    readonly end_lt: bigint;
    readonly gen_validator_list_hash_short: number;
    readonly gen_catchain_seqno: number;
    readonly min_ref_mc_seqno: number;
    readonly prev_key_block_seqno: number;
    readonly gen_software: GlobalVersion | undefined;
    readonly master_ref: BlkMasterInfo | undefined;
    readonly prev_ref: BlkPrevInfo;
    readonly prev_vert_ref: BlkPrevInfo | undefined;
}

// prev_blk_info$_ prev:ExtBlkRef = BlkPrevInfo 0;

// prev_blks_info$_ prev1:^ExtBlkRef prev2:^ExtBlkRef = BlkPrevInfo 1;

export type BlkPrevInfo = BlkPrevInfo_prev_blk_info | BlkPrevInfo_prev_blks_info;
// ...
export interface Block {
    readonly kind: 'Block';
    readonly global_id: number;
    readonly info: BlockInfo;
    readonly value_flow: ValueFlow;
    readonly state_update: Cell;
    readonly extra: BlockExtra;
}

/*
block_extra in_msg_descr:^InMsgDescr
  out_msg_descr:^OutMsgDescr
  account_blocks:^ShardAccountBlocks
  rand_seed:bits256
  created_by:bits256
  custom:(Maybe ^McBlockExtra) = BlockExtra;
*/

export interface BlockExtra {
    readonly kind: 'BlockExtra';
    readonly in_msg_descr: InMsgDescr;
    readonly out_msg_descr: OutMsgDescr;
    readonly account_blocks: ShardAccountBlocks;
    readonly rand_seed: Buffer;
    readonly created_by: Buffer;
    readonly custom: Maybe<McBlockExtra>;
}
// ...
export type ValueFlow = ValueFlow_value_flow | ValueFlow_value_flow_v2;

export interface ValueFlow_value_flow {
    readonly kind: 'ValueFlow_value_flow';
    readonly from_prev_blk: CurrencyCollection;
    readonly to_next_blk: CurrencyCollection;
    readonly imported: CurrencyCollection;
    readonly exported: CurrencyCollection;
    readonly fees_collected: CurrencyCollection;
    readonly fees_imported: CurrencyCollection;
    readonly recovered: CurrencyCollection;
    readonly created: CurrencyCollection;
    readonly minted: CurrencyCollection;
}

export interface ValueFlow_value_flow_v2 {
    readonly kind: 'ValueFlow_value_flow_v2';
    readonly from_prev_blk: CurrencyCollection;
    readonly to_next_blk: CurrencyCollection;
    readonly imported: CurrencyCollection;
    readonly exported: CurrencyCollection;
    readonly fees_collected: CurrencyCollection;
    readonly burned: CurrencyCollection;
    readonly fees_imported: CurrencyCollection;
    readonly recovered: CurrencyCollection;
    readonly created: CurrencyCollection;
    readonly minted: CurrencyCollection;
}
// ...
export interface McBlockExtra {
    readonly kind: 'McBlockExtra';
    readonly key_block: number;
    readonly shard_hashes: ShardHashes;
    readonly shard_fees: ShardFees;
    readonly prev_blk_signatures: Dictionary<number, CryptoSignaturePair>;
    readonly recover_create_msg: Maybe<InMsg>;
    readonly mint_msg: Maybe<InMsg>;
    readonly config: ConfigParams | undefined;
}
// ...
export interface GlobalVersion {
    readonly kind: 'GlobalVersion';
    readonly version: number;
    readonly capabilities: bigint;
}